python/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.8
Markdown==3.1.1
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
